--- daeyo-domain/domain-rds/src/main/java/com/joeun/domain/holding/entity/HoldingStatus.java
@@ -1,9 +1,0 @@
-package com.joeun.domain.holding.entity;
-
-public enum HoldingStatus {
-    OFFERED, // 홀딩 상태
-    FULFILLED, // 대여 확정
-    CANCELLED, // 사용자 홀딩 취소
-    EXPIRED, // 홀딩 시간 만료
-    REVOKED // 관리자 홀딩 철회
-}
