--- daeyo-domain/domain-rds/src/main/java/com/joeun/domain/holding/repository/HoldingRepository.java
@@ -1,7 +1,0 @@
-package com.joeun.domain.holding.repository;
-
-import com.joeun.domain.holding.entity.Holding;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HoldingRepository extends JpaRepository<Holding, Long> {
-}
