+++ daeyo-domain/domain-service/build.gradle
@@ -16,6 +16,9 @@ dependencies {
     api project(':daeyo-domain:domain-rds')
     api project(':daeyo-domain:domain-redis')
 
+    // Spring Data JPA
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
 }
