--- daeyo-domain/domain-service/src/main/java/com/joeun/service/holding/HoldingDomainService.java
@@ -1,14 +1,0 @@
-package com.joeun.service.holding;
-
-import com.joeun.domain.holding.service.HoldingRdsService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class HoldingDomainService {
-
-    private final HoldingRdsService holdingRdsService;
-
-
-}
